+++ src/app/calendar.tsx
@@ -8,37 +8,6 @@ export default function Calendar({startDate, semesterStart, weeks, deadlines}: {
   const [showCover, setshowCover] = useState(false)
   const [modalDeadlines, setModalDeadline] = useState<{date: Date, deadlines :(Deadline|undefined)[]}>({date: new Date(), deadlines: []})
 
-  const [showEditForm, setShowEditForm] = useState(false);
-  
-  const defaultDeadline: Deadline = {
-    name: "",
-    subject: "",
-    start: "",
-    due: "",
-    mark: 0,
-    room: "",
-    url: "",
-    info: "",
-    color: "1",
-  }
-
-  const [currentEdit, setCurrentEdit] = useState<Deadline>(defaultDeadline)
-  const [originalEdit, setOriginalEdit] = useState<Deadline>(defaultDeadline)
-
-  function handleEdit(stuff: Deadline){
-    setShowEditForm(true);
-    setCurrentEdit(stuff)
-    setOriginalEdit(stuff)
-  }
-
-  const handleChange = (e : React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    const { name, value } = e.target;
-    setCurrentEdit({
-      ...currentEdit,
-      [name]: value,
-    });
-  };
-
 
   function openModal(d: {date: Date, deadlines: (Deadline|undefined)[]}){
     setModalDeadline(d)
@@ -81,11 +50,9 @@ export default function Calendar({startDate, semesterStart, weeks, deadlines}: {
       <div className={`modalCover ${showModal && "active"} ${showCover&&"hidden"}`} onClick={()=>{closeModal()}} onScroll={()=>console.log("scroll")}>
         <div className="flex w-full h-full justify-center">
           <div className="bg-white dark:bg-slate-800 rounded-3xl modalCard">
-            <Modal semStart={semesterStart} deadlines={modalDeadlines.deadlines} today={modalDeadlines.date} showEditForm={showEditForm} setShowEditForm={setShowEditForm}currentEdit={currentEdit} handleChange={handleChange} setEditForm={setCurrentEdit} handleEdit={handleEdit} originalEditData={originalEdit}/>
+            <Modal semStart={semesterStart} deadlines={modalDeadlines.deadlines} today={modalDeadlines.date}/>
           </div>
         </div>
       </div>
-
-
   </div>
 }
