+++ src/app/modal.tsx
@@ -4,10 +4,40 @@ import { PiNotePencilBold } from "react-icons/pi"
 import DateIco from "./date"
 import DeadlineCard from "./deadlineCard"
 import EditForm from "./editForm"
-import { ChangeEventHandler } from "react"
+import { useState } from "react"
 
-export default function Modal({semStart, deadlines, today, currentEdit, handleChange, setEditForm, showEditForm, setShowEditForm, handleEdit, originalEditData} : {semStart: Date, deadlines: (Deadline | undefined)[], today:Date, currentEdit: Deadline, handleChange: ChangeEventHandler, setEditForm: Function, showEditForm: Boolean, setShowEditForm: Function, handleEdit: Function, originalEditData: Deadline}) {
+export default function Modal({semStart, deadlines, today} : {semStart: Date, deadlines: (Deadline | undefined)[], today:Date}) {
 
+  const [showEditForm, setShowEditForm] = useState(false);
+
+  const defaultDeadline: Deadline = {
+    name: "",
+    subject: "",
+    start: "",
+    due: "",
+    mark: 0,
+    room: "",
+    url: "",
+    info: "",
+    color: "1",
+  }
+  const [currentEdit, setCurrentEdit] = useState<Deadline>(defaultDeadline)
+  const [originalEdit, setOriginalEdit] = useState<Deadline>(defaultDeadline)
+
+
+  function handleEdit(stuff: Deadline){
+    setShowEditForm(true)
+    setCurrentEdit(stuff)
+    setOriginalEdit(stuff)
+  }
+
+  const handleChange = (e : React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
+    const { name, value } = e.target;
+    setCurrentEdit({
+      ...currentEdit,
+      [name]: value,
+    });
+  };
 
   //add all the deadline cards to the modal
   const deadlineRows = []
@@ -30,7 +60,7 @@ export default function Modal({semStart, deadlines, today, currentEdit, handleCh
       </div>
       
       <div className="h-[53vh] overflow-y-auto">
-        {showEditForm && <EditForm hide={()=>{setShowEditForm(false)}} day={today} originalData={originalEditData} data={currentEdit} handleChange={handleChange}/>}      
+        {showEditForm && <EditForm hide={()=>{setShowEditForm(false)}} day={today} originalData={originalEdit} data={currentEdit} handleChange={handleChange}/>}      
         {deadlineRows}
         {(deadlineRows.length == 0) && !showEditForm && (
       <div className="flex justify-center p-8">No Deadlines!!</div>)}
